Seeders + fiels 'slug' added in some models + minor fixes

+++ database/migrations/2026_05_06_154004_create_student_subject_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             // Связь с предметом
             $table->foreignId('subject_id')->constrained('subjects')->cascadeOnDelete();
 
-            $table->unsignedTinyInteger('grade')->nullable();
+            $table->decimal('grade', 3, 1)->nullable();
 
             $table->primary(['user_id', 'subject_id']);
         });
